+++ UVAtlas/isochart/packingcharts.cpp
@@ -821,7 +821,7 @@ namespace
                 XMStoreFloat3(&v2, vv2);
         
                 if (IsInZeroRange(v1.x) && IsInZeroRange(v2.x) 
-                    && fabs(v1.y) > 0.1f && fabs(v2.y) > 0.1f
+                    && fabsf(v1.y) > 0.1f && fabsf(v2.y) > 0.1f
                     && v1.y * v2.y < 0)
                 {
                     if (v1.y > v2.y)
@@ -835,7 +835,7 @@ namespace
                 }
 
                 if (IsInZeroRange(v1.y) && IsInZeroRange(v2.y) 
-                    && fabs(v1.x) > 0.1f && fabs(v2.x) > 0.1f
+                    && fabsf(v1.x) > 0.1f && fabsf(v2.x) > 0.1f
                     && v1.x * v2.x < 0)
                 {
                     if (v1.x < v2.x)
@@ -852,7 +852,7 @@ namespace
             }
              // pOrigin, pVertex1, pVertex2 in the same line,
              // need to move origin forward, and Calculate again
-             if (fabs(fZ) < ISOCHART_ZERO_EPS)
+             if (fabsf(fZ) < ISOCHART_ZERO_EPS)
              {
                 if (f1 < f2)
                 {
@@ -953,7 +953,7 @@ namespace
                 XMStoreFloat3(&v2, vv2);
 
                 if (IsInZeroRange(v1.x) && IsInZeroRange(v2.x)
-                    && fabs(v1.y) > 0.1f && fabs(v2.y) > 0.1f
+                    && fabsf(v1.y) > 0.1f && fabsf(v2.y) > 0.1f
                     && v1.y * v2.y < 0)
                 {
                     if (v1.y > v2.y)
@@ -967,7 +967,7 @@ namespace
                 }
 
                 if (IsInZeroRange(v1.y) && IsInZeroRange(v2.y)
-                    && fabs(v1.x) > 0.1f && fabs(v2.x) > 0.1f
+                    && fabsf(v1.x) > 0.1f && fabsf(v2.x) > 0.1f
                     && v1.x * v2.x < 0)
                 {
                     if (v1.x < v2.x)
@@ -984,7 +984,7 @@ namespace
             }
             // pOrigin, pVertex1, pVertex2 in the same line,
             // need to move origin forward, and Calculate again
-            if (fabs(fZ) < ISOCHART_ZERO_EPS)
+            if (fabsf(fZ) < ISOCHART_ZERO_EPS)
             {
                 if (f1 < f2)
                 {
@@ -1136,8 +1136,8 @@ namespace
             float r2 = VECTOR_ITEM(&border[ii]->uv, RadialAxis);
             float r3 = VECTOR_ITEM(&border[ii+1]->uv, RadialAxis);
 
-            if (fabs(t1-t2) < ISOCHART_ZERO_EPS
-            && fabs(t3-t2) < ISOCHART_ZERO_EPS)
+            if (fabsf(t1-t2) < ISOCHART_ZERO_EPS
+                && fabsf(t3-t2) < ISOCHART_ZERO_EPS)
             {
                 if((r1 >= r2 && r2 >= r3) || (r1 <= r2 && r2 <= r3))
                 {
@@ -2098,7 +2098,7 @@ static void BruteForceFoldChecking(
                         bFoundFold = true;
                         DPF(0, "Found fold in chart %zu...", ii);
                         DPF(0, "(%f, %f) (%f, %f) --> (%f, %f) (%f, %f)",
-                                v1.x, v1.y, v2.x, v2.y, v3.x, v3.y, v4.x, v4.y);
+                            double(v1.x), double(v1.y), double(v2.x), double(v2.y), double(v3.x), double(v3.y), double(v4.x), double(v4.y));
                     }
                 }
             }
@@ -2152,7 +2152,7 @@ static void BruteForceOverlappingChecking(
                         DPF(0, "Chart1 %zu, Chart2 %zu\n", ii, jj);
 
                         DPF(0, "(%f, %f) (%f, %f) --> (%f, %f) (%f, %f)",
-                            v1.x, v1.y, v2.x, v2.y, v3.x, v3.y, v4.x, v4.y);
+                            double(v1.x), double(v1.y), double(v2.x), double(v2.y), double(v3.x), double(v3.y), double(v4.x), double(v4.y));
 
                         assert(!bIsIntersect);
                     }
@@ -2210,9 +2210,9 @@ HRESULT CIsochartMesh::PackingCharts(
         }
     }
 
-    DPF(3, "Area lost rate is : %f", 1 - atlasInfo.fPackedChartArea / 
-           ((atlasInfo.fBoxRight-atlasInfo.fBoxLeft) * 
-           (atlasInfo.fBoxTop-atlasInfo.fBoxBottom)));
+    DPF(3, "Area lost rate is : %f", double(1 - atlasInfo.fPackedChartArea /
+        ((atlasInfo.fBoxRight - atlasInfo.fBoxLeft) *
+        (atlasInfo.fBoxTop - atlasInfo.fBoxBottom))));
 
     // 3. Normalize the atlas to [0.0, 1.0]
     NormalizeAtlas(chartList, atlasInfo);
@@ -2338,7 +2338,7 @@ HRESULT CIsochartMesh::PreparePacking(
             gutter);
 
     atlasInfo.fGutter = gutter * atlasInfo.fPixelLength;
-    DPF(2, "Pixel Length is %f", atlasInfo.fPixelLength);
+    DPF(2, "Pixel Length is %f", double(atlasInfo.fPixelLength));
     atlasInfo.fExpectedAtlasWidth = dwWidth * atlasInfo.fPixelLength;
     atlasInfo.fWidthHeightRatio = (static_cast<float>(dwWidth)) / (dwHeight);
     atlasInfo.fBoxTop = 0;
@@ -2743,14 +2743,14 @@ HRESULT CIsochartMesh::PackingOneChart(
 
     if (dwDirMinRotationId[dwPackDirection] == INVALID_INDEX)
     {
-        DPF(0, "2d area %f", pChart->m_fChart2DArea);	
-        DPF(0, "3d area %f", pChart->m_fChart3DArea);
-        DPF(0, "Face number %zu", pChart->m_dwFaceNumber);
-        DPF(0, "Vert number %zu", pChart->m_dwVertNumber);
+        DPF(0, "2d area %f", double(pChart->m_fChart2DArea));
+        DPF(0, "3d area %f", double(pChart->m_fChart3DArea));
+        DPF(0, "Face number %zu", double(pChart->m_dwFaceNumber));
+        DPF(0, "Vert number %zu", double(pChart->m_dwVertNumber));
 
         for (size_t ii=0; ii<pChart->m_dwVertNumber; ii++)
         {
-            DPF(0, "(%f, %f)", pChart->m_pVerts[ii].uv.x, pChart->m_pVerts[ii].uv.y);
+            DPF(0, "(%f, %f)", double(pChart->m_pVerts[ii].uv.x), double(pChart->m_pVerts[ii].uv.y));
         }
     }
      
@@ -3152,7 +3152,7 @@ HRESULT CIsochartMesh::CalculateChartBorders(
     if ((bFirstBorderOutside && !bSecondBorderInSide) 
         || (!bFirstBorderOutside && bSecondBorderInSide))
     {
-        DPF(1, "Dot value 1 = %f, Dot value 2 = %f", fDotValue1, fDotValue2);
+        DPF(1, "Dot value 1 = %f, Dot value 2 = %f", double(fDotValue1), double(fDotValue2));
         if (fabsf(fDotValue1) < 0.1f && fabsf(fDotValue2) > 0.9f)
         {
             bSecondBorderInSide = bFirstBorderOutside;
@@ -3396,8 +3396,8 @@ void CIsochartMesh::NormalizeAtlas(
             pVertex->uv.x = (pVertex->uv.x - atlasInfo.fBoxLeft) / fScaleW;
             pVertex->uv.y = (pVertex->uv.y - atlasInfo.fBoxBottom)/ fScaleH;
 
-            assert(_finite(pVertex->uv.x));
-            assert(_finite(pVertex->uv.y));
+            assert(_finite(double(pVertex->uv.x)));
+            assert(_finite(double(pVertex->uv.y)));
 
             if (pVertex->uv.x < 0.0f)
             {
@@ -3456,7 +3456,7 @@ void CIsochartMesh::OptimizeAtlasSignalStretch(
         pChart->m_fChart2DArea = pChart->CalculateChart2DArea();
         fTotal2DArea += pChart->m_fChart2DArea;
 
-        assert(_finite(pChart->m_fParamStretchL2) != 0);
+        assert(_finite(double(pChart->m_fParamStretchL2)) != 0);
 
         fTotal += IsochartSqrtf(
             (pChart->m_fParamStretchL2+ShiftError)*pChart->m_fChart2DArea);
@@ -3490,7 +3490,7 @@ void CIsochartMesh::OptimizeAtlasSignalStretch(
 
         pChart->ScaleChart(fScale);
 
-        assert(_finite(pChart->m_fParamStretchL2) != 0);
+        assert(_finite(double(pChart->m_fParamStretchL2)) != 0);
     }
 }
